apistack/requirements/base.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
astroid=1.6.1=py36_0
ca-certificates=2018.1.18=0
certifi=2018.1.18=py36_0
django=2.0.2=py36hd476221_0
django-environ==0.4.4=py36_0
djangorestframework=3.7.7=py_0
isort=4.3.4=py36_0
lazy-object-proxy=1.3.1=py36_0
mccabe=0.6.1=py36_0
ncurses=5.9=10
openssl=1.0.2n=0
pip=9.0.1=py36_1
pylint=1.8.2=py36_0
pymysql=0.8.0=py36_0
python=3.6.4=0
pytz=2018.3=py36_0
readline=7.0=0
setuptools=38.5.2=py36_0
six=1.11.0=py36_1
sqlite=3.20.1=2
tk=8.6.7=0
wheel=0.30.0=py36_2
wrapt=1.10.11=py36_0
xz=5.2.3=0
zlib=1.2.11=0
